requirements.txt
streamlit
pandas
numpy
plotly.express
plotly.graph_objects
streamlit.components.v1